+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* `company-semantic-insert-arguments`: New option. Like in `company-clang`.
 * `company-semantic-begin-after-member-access`: New option. Similar to the one
   in `company-clang`.
 * `company-capf` accepts `:company-prefix-length` property function.